+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - `Moar.Opts.get/3` and `Moar.Opts.take/2` accept keyword lists with a mix of "valueless" keys and regular keys, like
   `[:a, b: 2]`, where the default value for a "valueless" key is `true`.
 - Add `Moar.Opts.pop/3` which pops an opt out of an opts enum.
+- Add `:apply` and `:map` options to `assert_eq/3` to run one or more functions on `left` or `right` (`:apply`) or run one
+  or more functions on each element in `left` or `right` (`:map`).
 
 ## 1.50.0